rename folder so it is not ignored by git
